+++ {{cookiecutter.project_name}}/src/main/pages/base_serializer.py
@@ -47,6 +47,10 @@ class BasePageSerializer(serializers.ModelSerializer):
         if not request:
             return None
 
+        in_preview_panel = getattr(request, "in_preview_panel", False)
+        if in_preview_panel:
+            return None
+
         if not hasattr(request, "user"):
             return None